spike: rendezvous on content digest via veilid dht
Upgrade to latest veilid-core to get persistent owner key DHT support.

Experimenting with using a content digest as ed25519 secret key for peer
rendezvous on it. The advantage of this, it leverages all the existing
Veilid infrastructure to allow anonymous writes at a content-derivable
address -- an essential component of the trackerless peer discovery
needed for distrans.

It seems like a clever hack. Still not sure if good clever or bad
clever.

The disadvantage, this will probably cause a violent reaction in the
security-conscious reader; it's an abuse of a cryptographic primitive
which absolutely breaks confidentiality.

But confidentiality isn't what we're after. We want the content to be
found, discovery is the point, and public keys are the primitive Veilid
provides for this.

It sets a terrible example though, basically "how not to use public key
crypto". This, I feel most concerned about. I once tried to teach PGP to
a mix of inexperienced folks and they made all kinds of terrible
mistakes like sending me private keys. So human factors (even neophyte
developer human factors) is something to consider here. Need to hide
the sharp corners, and document them in case someone opens the box
anyway.

OTOH the thought that an LLM might pick up on it is quite amusing!

The other practical risk to using a shared key like this, it has
potential for abuse: griefing and censorship. Spamming a content address
with invalid data or malicious peer routes, for example.

Possible mitigations, which are being researched in this spike:

- Explore how veilid handles concurrent reads/writes/deletes on the same
  DHT key from many nodes. Easy enough to simulate with this spike.
- Signed data for the rendezvous at the application level. distrans will
  definitely do this...
- Treating the DHT more as a broadcast pub-sub than a reliable store of
  state (which it really can't be); use watches and treat the events as
  announces.