feat: havemap and peermap, scalable piece size.
A havemap is a DHT key that a peer publishes, representing the pieces
that it has, as a bitmap chunked over 32k subkey values.

A peermap is another DHT key that a peer publishes, representing the
other peers that a peer knows about.

The piece size is now scalable, so that an index and havemap can handle
larger files. It's also a potential knob to tune throughput under
different conditions (reliability of the Veilid network, trustworthiness
of peers in a swarm, etc). The maximum piece size that can be expressed
by the protocol is now 2GiB. If distrans shares go far into exabyte scale
then I think we can add more extended index DHTs ;)

It remains to be seen whether this will be a long-term viable solution
to all large file scaling needs. At a certain point, it might be
necessary to cap the piece size and split large indexes over multiple
DHT keys.