chore!: rebrand this project as stigmerge
Renaming distrans to stigmerge for a few reasons:

It's not original; distrans is a secret messaging system from Frank
Herbert's Dune series. While I borrowed it as someone deeply inspired by
Herbert's works, it's not my name. "IP ownership" changes hands over
time, and who knows if trademark could be used to censor this work. Why
even risk it?

Potential connotations with "de-transitioning" based on community
feedback. While this was never the intent, connotations and cultural
context matter. Especially now, the trans community needs our support
and respect. I'm happy to rename it because the last thing I want is to
cause more pain in this world. When I see horrible people doing horrible
things and making "excuses" for it these days (you know who I'm
referring to), well, I wanna do exactly the opposite of that.

So what's the deal with stigmerge? Isn't that another weird awkward
funny word? Well, maybe, but here's why it's sticking in my head:

From wikipedia:

> The term "stigmergy" was introduced by French biologist Pierre-Paul Grassé in 1959 to refer to termite behavior. He defined it as: "Stimulation of workers by the performance they have achieved."
> ... and captures the notion that an agent’s actions leave signs in the environment, signs that it and other agents sense and that determine and incite their subsequent actions.

The very purpose of this software, what it does, is to distribute
content through swarms of peers running this software, exchanging pieces
of it amongst themselves, cooperating.

And "leaving signs in the environment" is pretty close to how the DHT is
or will be used to exchange private routes, indexes, block maps and
neighbors.

Peers are "merging" pieces of the content together. So "stigmerge" is
kind of two overlapping words.

I've already claimed the crate names. Going to sleep on it but I think
this is the name I'll likely go with.

And then there's how it might be used in conversation. Do I "stigmerge"
a file to get it? To send it? Both? Does "stig" become a shorthand for
it?

I suppose the only confusing connotation I can think of with this new
name are STIGs which are a pretty different thing. So you can still stig
me some STIGs while you stig.